Fixed lots of loose ends
VERTICAL and HORIZONAL is better than vertical and horizontal for constants because the lowercase words constantly collides with regular variable names, functions etc. It got annoying.

Changed the config_widget!() function to use `k in fieldnames(typeof(w))` That means it will work for a wider variety of Widget subclasses. E.g. a PushButton has a text field but not a CustomWidget, unless you add it to the property list.