Make Assoc an actual dictonary
Motivation: I needed eltype(), valtype() and keytype() functions and started implementing them myself, only to realize that if you inherent the abstract type AbstractDict, you get all this stuff for free.

A lot of dictionary functions are implemented for AbstractDict rather than the concrete types. The functionality is built on top of primitives defined for the concrete types.